requirements.txt
aiohttp>=3.9
aiosqlite>=0.20
python-dotenv>=1.0
telethon>=1.34
pysocks>=1.7
google-generativeai>=0.8
web3>=6.20
